# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gaming Zone Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  admin: "admin",
  staff: "staff",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: RegisterRequestRole;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type DeviceType = (typeof DeviceType)[keyof typeof DeviceType];

export const DeviceType = {
  PS5: "PS5",
  PS4: "PS4",
  PC: "PC",
} as const;

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  available: "available",
  in_use: "in_use",
} as const;

export interface ActiveSessionInfo {
  sessionId: string;
  startTime: string;
}

export interface Device {
  id: string;
  name: string;
  type: DeviceType;
  hourlyRate: number;
  status: DeviceStatus;
  createdAt: string;
  activeSession?: ActiveSessionInfo;
}

export type CreateDeviceRequestType =
  (typeof CreateDeviceRequestType)[keyof typeof CreateDeviceRequestType];

export const CreateDeviceRequestType = {
  PS5: "PS5",
  PS4: "PS4",
  PC: "PC",
} as const;

export interface CreateDeviceRequest {
  name: string;
  type: CreateDeviceRequestType;
  hourlyRate: number;
}

export type UpdateDeviceRequestType =
  (typeof UpdateDeviceRequestType)[keyof typeof UpdateDeviceRequestType];

export const UpdateDeviceRequestType = {
  PS5: "PS5",
  PS4: "PS4",
  PC: "PC",
} as const;

export interface UpdateDeviceRequest {
  name?: string;
  type?: UpdateDeviceRequestType;
  hourlyRate?: number;
}

export interface StartSessionRequest {
  deviceId: string;
}

export interface StopSessionRequest {
  deviceId: string;
}

export interface Session {
  id: string;
  deviceId: string;
  deviceName: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  hourlyRate: number;
  /** @nullable */
  totalAmount?: number | null;
  createdBy: string;
  operatorName: string;
  createdAt: string;
}

export interface SessionListResponse {
  sessions: Session[];
  total: number;
  page: number;
  totalPages: number;
}

export interface DailyRevenuePoint {
  date: string;
  revenue: number;
  sessions: number;
}

export interface DeviceUsageStat {
  deviceName: string;
  sessions: number;
  revenue: number;
}

export interface DailyReportResponse {
  dailyRevenue: DailyRevenuePoint[];
  deviceUsage: DeviceUsageStat[];
  todayEarnings: number;
  todaySessions: number;
  /** @nullable */
  mostUsedDevice: string | null;
}

export interface MonthlyReportResponse {
  monthlyEarnings: number;
  monthlySessions: number;
  avgDailyRevenue: number;
}

export interface Settings {
  id: string;
  defaultHourlyRate: number;
  currency: string;
  businessName: string;
  timezone: string;
}

export interface UpdateSettingsRequest {
  defaultHourlyRate?: number;
  currency?: string;
  businessName?: string;
  timezone?: string;
}

export type GetSessionsParams = {
  page?: number;
  limit?: number;
  deviceName?: string;
  startDate?: string;
  endDate?: string;
  filter?: GetSessionsFilter;
};

export type GetSessionsFilter =
  (typeof GetSessionsFilter)[keyof typeof GetSessionsFilter];

export const GetSessionsFilter = {
  today: "today",
  week: "week",
  all: "all",
} as const;

export type GetDailyReportParams = {
  days?: number;
};