# lib/api-zod/src/generated/types/createDeviceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gaming Zone Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDeviceRequestType } from "./createDeviceRequestType";

export interface CreateDeviceRequest {
  name: string;
  type: CreateDeviceRequestType;
  hourlyRate: number;
}