# lib/api-zod/src/generated/types/dailyReportResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gaming Zone Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyRevenuePoint } from "./dailyRevenuePoint";
import type { DeviceUsageStat } from "./deviceUsageStat";

export interface DailyReportResponse {
  dailyRevenue: DailyRevenuePoint[];
  deviceUsage: DeviceUsageStat[];
  todayEarnings: number;
  todaySessions: number;
  /** @nullable */
  mostUsedDevice: string | null;
}