# lib/api-zod/src/generated/types/sessionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gaming Zone Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from "./session";

export interface SessionListResponse {
  sessions: Session[];
  total: number;
  page: number;
  totalPages: number;
}