# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gaming Zone Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: string;
  defaultHourlyRate: number;
  currency: string;
  businessName: string;
  timezone: string;
}