# lib/api-zod/src/generated/types/updateSettingsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gaming Zone Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSettingsRequest {
  defaultHourlyRate?: number;
  currency?: string;
  businessName?: string;
  timezone?: string;
}